Replace external with internal link
Replace the external link for freecatphotoapp.com with an internal link that references the footer on the current webpage.  Remove the target attribute from the anchor tag so that the link will open in the same window or tab.  Add a comment describing what I have learned in this lesson.