Set box size using concise padding notation
Set the size of the colored boxes by adjusting the padding values in a single line rather than using the four padding properties for each individual side to remove excess lines of code.  Add a comment describing what I learned in this lesson about this more concise notation for padding.