Reduce image size for small screens
Reduce the size of the penguin image for screens less than 350px in width using a media query to redefine the value of the variable for the image size.  Add a comment explaining what I learned in this lesson about how a media query can be used to change a variable's value.