Adjust size and grow/shrink values for color boxes
Adjust the flex-grow, flex-shrink, and flex-basis values for the color boxes so that the boxes have the same size for a 300px wide container, but the blue box grows twice as fast as the red box for container widths greater than 300px and shrinks twice as fast as the red box for container widths of less than 300px.  Also add a comment explaining the "shorthand" notation used to combine flex-grow, flex-shrink, and flex-basis values into a single line.