Create dynamic columns with auto-fit
Create dynamic columns in the second grid using the auto-fit value for their width.  Thus, items will be arranged in as many adjacent columns as the screen can accomodate until all items are in the same row, and then for wider screens columns will be stretched to fill the full width of the screen.  Also, add a comment explaining what this lesson taught me about the auto-fill value for the repeat function.