Initialize three new variables
Initialize three new variables.  Variables a and b have numerical values so that mathematical operations can be performed on them.  Variable c is a string, and so can be concatenated with other strings later.  Add a comment describing what this lesson teaches about initialization and the data types "undefined", "number", and "string".