Create shopping list as multi-dimensional array
Create a multi-dimensional array to store my shopping list.  Every item in the array is an array containing first the item name as a string and then the number of that item that I need to purchase as a number.  Add a comment explaining what I learned in this lesson.