Create function to shift and push array values
Create a function that removes a value from the first position in an array, adds the number specified in the arguments to the end of the array, and returns the manipulated array as well as the value that was removed from the function.  Add a comment explaining what I learned about writing functions in this lesson.