Create function to find values from 25 to 50
Create a function that determines whether a number is greater than or equal to 25 and is also less than or equal to 50, and then returns a statement of "Yes" if the value is in the desired range and "No" if it is not.  Also, add a comment explaining how the 'logical and' operator works.