Create function to specify numerical range
Create a function that determines whether the argument is less than 5, 10, 15, 20, or 25, or is 25 or bigger, and returns a statement indicating the results.  Add a comment explaining how if, else if, and else statements can be chained together in order to provide a larger number of possible outcomes for a function.