Create function to lookup and return profile properties
Create a function that iterates through contacts in an array in order to determine if the contact listed in the arguments exists and whether it has the property specified in the arguments.  If the contact exists and has the property, the function returns the value of the property.  Otherwise, the function returns a statement that the contact or property does not exist as appropriate.  Add a comment explaining what I learned in this lesson by applying skills with for loops and if/else if statements.