Edit function so that second declaration of i does not override first
Edit the function checkScope so that the second declaration of the variable i does not override the first declaration, and instead creates a second variable with a scope local to the if statement.  Add a comment explaining how the "let" keyword determines scope for variables declared within functions, blocks, statements, or expressions.