Create function to square positive integers only
Create a function that filters the contents of an array and selects only the positive integers, and then squares those entries and returns an array of the results.  Add a comment explaining what I learned in this lesson about using arrow notation with higher order functions such as map() and filter(), and about the higher order functions themselves.