Simplify code using destructuring assignment for arguments
Simplify the code for the function half by using destructuring assignment to only pass the needed min and max properties from the object stats into the function.  This allows the code to be shorter and cleaner.  Add a comment explaining how destructuring assignment can be used to pass arguments to a function.