Create Vegetable constructor function with name property
Create a Vegetable constructor function so that the objects created have the property name with the value equal to the argument of the Vegetable function.  Add a comment explaining what I learned in this lesson about the new syntax using the class keyword for defining constructor functions available in ES6.