Create constructor function with getters, setters converting Celsius/Farenheit
Create a constructor function that takes a temperature in Farenheit, and then when using a getter returns the value in Celsius, and when using a setter takes a value in Celsius and converts it to Farenheit before setting the object value.  Add a comment explaining how getters and setters work.